comp/core/workloadmeta: move the runtime property merge to sbomutil
The merge that carries LastSeenRunning, HasSetSuidBit and RunningAsRoot
from a system-probe report onto a Trivy SBOM sits in the remote SBOM
collector, its only caller today. It is a pure function of two CycloneDX
documents, so it moves to sbomutil, beside the compression helpers, where
the sbom check can reach it too. The property names and the tests move
with it.

Behavior is unchanged.
